Add download and upload CLI options

+++ test/cli/deployExperiments.cjs
@@ -9,12 +9,39 @@ const CLIParser = require('../shared/CLIParser.cjs');
 const path = require('path');
 const child_process = require('child_process');
 
-// Get branch
-let branch = CLIParser.parseOption({env: 'GITHUB_REF', cli: 'branch'});
-console.log('[deployExperiments.cjs] branch is ' + branch);
+// Get download CLI option (download psyexp files from stager?)
+let download = CLIParser.parseOption({cli: 'download'}, false);
+download = download !== undefined;
+console.log('[compileExperiments.cjs] download is ' + download);
 
-// Base URL to experiment on Stager
-let baseURL = 'https://staging.psychopy.org/experiments/html/' + branch;
+// Get upload CLI option (upload psyexp files compiled to JS to stager?)
+let upload = CLIParser.parseOption({cli: 'upload'}, false);
+upload = upload !== undefined;
+console.log('[compileExperiments.cjs] upload is ' + upload);
+
+// If upload enabled, get branch CLI option
+let branch;
+if (upload) {
+  branch = CLIParser.parseOption({env: 'GITHUB_REF', cli: 'branch'});
+  console.log('[deployExperiments.cjs] branch is ' + branch);
+}
+
+
+// Get url CLI option and construct baseUrl
+const url = CLIParser.parseOption({cli: 'url'}, false);
+let baseUrl;
+if (url !== undefined) {
+  // url defined? Use it as baseUrl, else check upload and branch settings
+  baseUrl = url;
+} else if (upload) {
+  // url not defined but upload enabled? Construct baserUrl from Stager and branch
+  baseUrl = 'https://staging.psychopy.org/experiments/html/' + branch;
+} else {
+  // Could not construct baseUrl
+  throw "[wdio.conf.cjs] Could not construct baseUrl, because url was not specified and upload was false";
+}
+console.log('[wdio.conf.cjs] baseUrl is ' + baseUrl);
+
 
 // get files in dirPath and each of its subdirectories
 // https://coderrocketfuel.com/article/recursively-list-all-the-files-in-a-directory-using-node-js
@@ -57,33 +84,35 @@ const readDirSyncRecursive = (fromPath, toPath) => {
 
 // Download experiments to temporary directory
 (async () => {
-  // Delete experiments directory (it not empty)
-  console.log('[deployExperiments.cjs] cleaning up temporary folders');
-  try {
-    fs.rmdirSync(Paths.dir_experiments, {recursive: true});
-  } catch (e) {}
-  // Recreate experiments directory (with workaround for Windows)
-  let dirCreated = false;
-  let startTime = new Date().getTime();
-  while (!dirCreated) {
+  if (download) {
+    // Delete experiments directory (it not empty)
+    console.log('[deployExperiments.cjs] cleaning up temporary folders');
     try {
-      fs.mkdirSync(Paths.dir_experiments);
-      dirCreated = true;
-    } catch (e) {
-      // 10 seconds passed? Throw error
-      if ((new Date().getTime()) - startTime > 10000) {
-        throw (e);
+      fs.rmdirSync(Paths.dir_experiments, {recursive: true});
+    } catch (e) {}
+    // Recreate experiments directory (with workaround for Windows)
+    let dirCreated = false;
+    let startTime = new Date().getTime();
+    while (!dirCreated) {
+      try {
+        fs.mkdirSync(Paths.dir_experiments);
+        dirCreated = true;
+      } catch (e) {
+        // 10 seconds passed? Throw error
+        if ((new Date().getTime()) - startTime > 10000) {
+          throw (e);
+        }
       }
     }
-  }
 
-  // Download experiments
-  console.log('[deployExperiments.cjs] downloading experiments');
-  downloadResults = await Stager.ftpRequest((client, basePath) => {
-    return client.downloadDir(basePath + '/experiments/js', Paths.dir_experiments);
-  }, true);
+    // Download experiments
+    console.log('[deployExperiments.cjs] downloading experiments');
+    downloadResults = await Stager.ftpRequest((client, basePath) => {
+      return client.downloadDir(basePath + '/experiments/js', Paths.dir_experiments);
+    }, true);
+  }
   
-  // List of  experiments
+  // List of experiments
   let experiments = fs.readdirSync(Paths.dir_experiments);
   console.log('[deployExperiments.cjs] preparing ' + experiments.length + ' experiments');
   // Get template
@@ -113,16 +142,21 @@ const readDirSyncRecursive = (fromPath, toPath) => {
       resources = [];
     }
     fs.writeFileSync(
-      Paths.dir_experiments + '/' + experiment + '/resources.cjson', 
+      Paths.dir_experiments + '/' + experiment + '/resources.json', 
       JSON.stringify({
         resources: resources,
-        resourceDirectory: baseURL + '/' + experiment + '/resources/'
+        resourceDirectory: baseUrl + '/' + experiment + '/resources/'
       })
     );    
 
     // Copy dist/ to lib/
-    console.log('[deployExperiments.cjs] copying lib');
-    fs.mkdirSync(Paths.dir_experiments + '/' + experiment + '/lib');
+    if (fs.existsSync(Paths.dir_experiments + '/' + experiment + '/lib')) {
+      console.log('[deployExperiments.cjs] lib directory already exists');
+    } else {
+      console.log('[deployExperiments.cjs] creating lib directory');
+      fs.mkdirSync(Paths.dir_experiments + '/' + experiment + '/lib');
+    }
+    console.log('[deployExperiments.cjs] copying lib');    
     includes.map((include) => { 
       fs.copyFileSync(
         './dist/' + include,
@@ -132,8 +166,11 @@ const readDirSyncRecursive = (fromPath, toPath) => {
   }
 
   // Upload compiled experiments to stager
-  Stager.uploadDirectory(
-    Paths.dir_experiments,
-    'experiments/html/' + branch
-  )
+  if (upload) {
+    console.log('[deployExperiments.cjs] uploading experiments');
+    Stager.uploadDirectory(
+      Paths.dir_experiments,
+      'experiments/html/' + branch
+    )
+  }
 })();
